# PERF: DataFrame[object].fillna